I removed the function process_slide from the main to a new module
